Use autoboxing between primitive types and Object wrappers
Also, better string concatenation.

+++ compatibility-server/src/main/java/com/vaadin/v7/ui/ProgressBar.java
@@ -57,7 +57,7 @@ public class ProgressBar extends AbstractField<Float>
      *            the initial progress value
      */
     public ProgressBar(float progress) {
-        setValue(Float.valueOf(progress));
+        setValue(progress);
     }
 
     /**
@@ -154,7 +154,7 @@ public class ProgressBar extends AbstractField<Float>
     protected void setInternalValue(Float newValue) {
         super.setInternalValue(newValue);
         if (newValue == null) {
-            newValue = Float.valueOf(0);
+            newValue = 0F;
         }
         getState().state = newValue;
     }
