+++ uitest/src/main/java/com/vaadin/tests/application/VaadinSessionAttribute.java
@@ -28,8 +28,8 @@ public class VaadinSessionAttribute extends AbstractReindeerTestUI {
 
     @Override
     protected void setup(VaadinRequest request) {
-        getSession().setAttribute(ATTR_NAME, Integer.valueOf(42));
-        getSession().setAttribute(Integer.class, Integer.valueOf(42 * 2));
+        getSession().setAttribute(ATTR_NAME, 42);
+        getSession().setAttribute(Integer.class, 42 * 2);
 
         addComponent(
                 new Button("Show attribute values", new Button.ClickListener() {
@@ -52,7 +52,7 @@ public class VaadinSessionAttribute extends AbstractReindeerTestUI {
 
     @Override
     protected Integer getTicketNumber() {
-        return Integer.valueOf(9514);
+        return 9514;
     }
 
 }
